+++ CHANGELOG.md
@@ -8,6 +8,7 @@ v5.0 (currently in beta)
 * Added `JsonBuffer::strdup()` to make a copy of a string (issues #10, #57)
 * Implicitly call `strdup()` for `String` but not for `char*` (issues #84, #87)
 * Added support of non standard JSON input (issue #44)
+* Added support of comments in JSON input (issue #88)
 * Redesigned `JsonVariant` to leverage converting constructors instead of assignment operators (issue #66)
 * Switched to new the library layout (requires Arduino 1.0.6 or above)